Cleaned up code, reorganized files
