+++ workers-rights-app/.gitignore
@@ -10,6 +10,7 @@ npm-debug.*
 web-build/
 web-report/
 *MyApiKeys.js
+*MyTwitterKeys.js
 
 # macOS
 .DS_Store
