--- workers-rights-app/screens/Footer.js
@@ -1,62 +1,0 @@
-import React, {Component, PropTypes} from 'react';
-import {Button, StyleSheet, TextInput, View} from 'react-native';
-
-const styles = StyleSheet.create({
-  footer: {
-    height: 48,
-    flexDirection: 'row',
-  },
-  input: {
-    flex: 1,
-  },
-});
-
-class Footer extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      text: 'testing',
-    };
-  }
-  render() {
-    return (
-      <View style={styles.footer} >
-        <TextInput
-          maxLength={140}
-          onChangeText={(text) => {this.setState({text});}}
-          style={styles.input}
-          value={this.state.text}
-        />
-         <Button
-          onPress={() => {
-            var myPromise = this.props.twitter.rest.get('statuses/user_timeline', {screen_name: 'NLRB', count: 3})
-              .then(console.log("done!"));
-            myPromise.then((info) => {
-              for(var x of info) {
-                console.log(x.text, "\n\n");
-              }
-              //console.log("text of first object: ", info[0].text)
-            });
-            this.setState({text: ''});
-            //console.log(myPromise);
-          }}
-          title="submit"
-        />
-        {/* <Button
-          onPress={() => {
-            this.props.twitter.rest.post('statuses/update', {status: this.state.text})
-              .catch(console.warn);
-            this.setState({text: ''});
-          }}
-          title="submit"
-        /> */}
-      </View>
-    );
-  }
-}
-
-// Footer.propTypes = {
-//   twitter: PropTypes.object.isRequired,
-// };
-
-export default Footer;
